Move caps protocol to hearth-types
